chore:  massive work push

+++ src/application/handlers/bankAccount/commandHandler/delete-account-command.handler.ts
@@ -7,6 +7,8 @@ import {
   type IBankAccountRepository,
 } from '../../../ports/bank-account-repository.interface';
 import { CommandResponse } from '../../../../shared/generic-class/command-response.class';
+import { ErrorMessageConst } from '../../../../shared/consts/error.const';
+import { Role } from '../../../../shared/consts/role.const';
 
 @CommandHandler(DeleteAccountCommand)
 export class DeleteAccountCommandHandler
@@ -22,6 +24,19 @@ export class DeleteAccountCommandHandler
     command: DeleteAccountCommand,
   ): Promise<CommandResponse<boolean>> {
     try {
+      const securityContext = this.securityContextProvider.getSecurityContext();
+      const account = await this.bankAccountRepository.getItem({
+        id: command.id,
+      });
+      if (!account) {
+        return CommandResponse.failure(ErrorMessageConst.CONTENT_NOT_FOUND);
+      }
+      if (
+        !securityContext.roles.includes(Role.Admin) ||
+        !account.idsAllowedToDelete?.includes(securityContext.userId)
+      ) {
+        return CommandResponse.failure(ErrorMessageConst.FORBIDDEN);
+      }
       const response = await this.bankAccountRepository.deleteAccount(
         command.id,
       );
